cv-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
onnxruntime==1.22.0
onnx==1.17.0
opencv-python-headless==4.11.0.86
pyvips==3.1.1
minio==7.2.16
python-ulid==3.1.0
python-multipart==0.0.20
pydantic-settings==2.9.1
httpx==0.28.1
numpy==2.2.6
Pillow==11.2.1
pyyaml==6.0.2
cryptography==46.0.0
